Script for downloading images
this python script downloads images from: https://skyserver.sdss.org/dr18/SkyServerWS/ImgCutout/getjpeg URL
which on adding the location data in RA and DEC format gives the image of space at the location from SDSS database.
it also add the path of the image to its corresponding row under a new column "image file path". later to be used for building the machine learning model. the scale , height and width of the images are chosen by physical observation of images to ensure visibility of the particular celestial object.